@@ -3,13 +3,17 @@
     "description": "A library to dynamically sort collections / queries",
     "type": "library",
     "require": {
-        "symfony/http-foundation": "^2.8|^3.0|^4.0|^5.0"
+        "php": ">=8.1",
+        "symfony/http-foundation": "^5.4|^6.0|^7.0"
     },
     "require-dev": {
-        "phpspec/phpspec": "^2.5",
-        "doctrine/orm": "^2.5",
-        "symfony/property-access": "^2.8|^3.0|^4.0|^5.0",
-        "twig/twig": "^1.24"
+        "doctrine/orm": "^2.5|^3.0",
+        "symfony/property-access": "^5.4|^6.0|^7.0",
+        "twig/twig": "^3.0.0",
+        "vimeo/psalm": "^6.9",
+        "phpunit/phpunit": "^10.5.40",
+        "dg/bypass-finals": "^1.9",
+        "friendsofphp/php-cs-fixer": "^3.72"
     },
     "autoload": {
         "psr-4": {
@@ -20,11 +24,6 @@
         "doctrine/orm": "Needed for use of ORMApplier",
         "symfony/property-access": "Needed for use of ArrayAdapter"
     },
-    "autoload-dev": {
-        "psr-4": {
-            "spec\\UnZeroUn\\Sorter\\": "spec"
-        }
-    },
     "license": "MIT",
     "authors": [
         {
@@ -34,7 +33,7 @@
     ],
     "extra": {
         "branch-alias": {
-            "dev-master": "0.2.x-dev"
+            "dev-main": "1.0.x-dev"
         }
     }
 }
